Move DLLEXPORT define include file
The include file needs to be used by the C interface.
