+++ SlashGaming-Diablo-II-API/include/cxx/game_address_locator/game_address_locator_interface.hpp
@@ -43,7 +43,7 @@
 #include <memory>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -119,5 +119,5 @@ SGD2MAPI_GameAddressLocatorInterface_Destroy(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_ADDRESS_LOCATOR_GAME_ADDRESS_LOCATOR_INTERFACE_HPP_
