+++ SlashGaming-Diablo-II-API/include/cxx/game_address_locator/game_offset.hpp
@@ -47,7 +47,7 @@
 #include <memory>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -189,5 +189,5 @@ SGD2MAPI_GameOffset_GetOffset(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_ADDRESS_LOCATOR_GAME_OFFSET_HPP_
