+++ SlashGaming-Diablo-II-API/include/cxx/game_address_locator/game_ordinal.hpp
@@ -44,7 +44,7 @@
 #include <cstdint>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -185,5 +185,5 @@ SGD2MAPI_GameOrdinal_GetOrdinal(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_ADDRESS_LOCATOR_GAME_ORDINAL_HPP_
