+++ SlashGaming-Diablo-II-API/include/cxx/game_patch/game_branch_patch.hpp
@@ -50,7 +50,7 @@
 #include <vector>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -314,5 +314,5 @@ SGD2MAPI_GameBranchPatch_Remove(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_PATCH_GAME_BRANCH_PATCH_HPP_
