+++ SlashGaming-Diablo-II-API/include/cxx/game_patch/game_buffer_patch.hpp
@@ -51,7 +51,7 @@
 #include <vector>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -259,5 +259,5 @@ SGD2MAPI_GameBufferPatch_Remove(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_PATCH_GAME_BUFFER_PATCH_HPP_
