+++ SlashGaming-Diablo-II-API/include/cxx/game_patch/game_nop_patch.hpp
@@ -48,7 +48,7 @@
 #include <memory>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -224,5 +224,5 @@ SGD2MAPI_GameNopPatch_Remove(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_PATCH_GAME_NOP_PATCH_HPP_
