+++ SlashGaming-Diablo-II-API/include/cxx/game_patch/game_patch_interface.hpp
@@ -42,7 +42,7 @@
 #include <memory>
 #endif // __cplusplus
 
-#include "../dllexport_define.inc"
+#include "../../dllexport_define.inc"
 
 #ifdef __cplusplus
 namespace sgd2mapi {
@@ -142,5 +142,5 @@ SGD2MAPI_GamePatchInterface_Remove(
 
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
-#include "../dllexport_undefine.inc"
+#include "../../dllexport_undefine.inc"
 #endif // SGD2MAPI_CXX_GAME_PATCH_GAME_PATCH_INTERFACE_HPP_
