+++ SlashGaming-Diablo-II-API/include/dllexport_define.inc
